Send submit character in separate write to simulate human input
Based on https://dr-knz.net/bubbletea-control-inversion.html

A future iteration of this should be to individually send "keystrokes",
that is, for e.g. choosing the left confirmation button in a confirm,
first send the left arrow key, then the submit button. If multiple are
required, do it for multiple. I think this mostly is required for
control sequences, not so much text.

Refs #3